+++ platform/ios/src/Mapbox.h
@@ -33,6 +33,7 @@ FOUNDATION_EXPORT MLN_EXPORT const unsigned char MapboxVersionString[];
 #import "MLNHillshadeStyleLayer.h"
 #import "MLNImageSource.h"
 #import "MLNLight.h"
+#import "MLNSky.h"
 #import "MLNLineStyleLayer.h"
 #import "MLNLocationManager.h"
 #import "MLNLoggingConfiguration.h"